rename, feat: 기존 FolderPage -> SharedFolderPage, FolderPage 새로 추가
