+++ src/Main.jsx
@@ -3,6 +3,7 @@ import App from "./components/App";
 import LandingPage from "./pages/LandingPage";
 import SigninPage from "./pages/SigninPage";
 import SignupPage from "./pages/SignupPage";
+import SharedFolderPage from "./pages/SharedFolderPage";
 import FolderPage from "./pages/FolderPage";
 import NotFoundPage from "./pages/NotFoundPage";
 
@@ -14,9 +15,8 @@ function Main() {
           <Route index element={<LandingPage />} />
           <Route path="signin" element={<SigninPage />} />
           <Route path="signup" element={<SignupPage />} />
-          <Route path="folder">
-            <Route index element={<FolderPage />} />
-          </Route>
+          <Route path="shared" element={<SharedFolderPage />} />
+          <Route path="folder" element={<FolderPage />} />
           <Route path="*" element={<NotFoundPage />} />
         </Route>
       </Routes>
