+++ src/components/SharedFolderPage/Hero.css
