+++ src/components/SharedFolderPage/SearchBar.css
