+++ src/pages/SharedFolderPage.css
@@ -1,8 +1,8 @@
 /*********************
-      FolderPage
+   SharedFolderPage
 *********************/
 
-.FolderPage-container {
+.SharedFolderPage-container {
   width: 325px;
   padding: 20px 0;
 }
@@ -12,14 +12,14 @@
 *********************/
 
 @media (min-width: 768px) {
-  .FolderPage-container {
+  .SharedFolderPage-container {
     width: 704px;
     padding: 40px 0;
   }
 }
 
 @media (min-width: 1124px) {
-  .FolderPage-container {
+  .SharedFolderPage-container {
     width: 1060px;
   }
 }
